exchange data provider moved to parent class
